+++ lib/knossoscli/cli.rb
@@ -1,4 +1,9 @@
+require 'knossos'
 require 'thor'
 
 class KnossosCLI::CLI < Thor
+  desc 'carve', 'Carve a maze'
+  def carve
+    puts Knossos::BinaryTree.carve(Knossos::Grid.new())
+  end
 end